% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_checks.R
\name{check_consec_na}
\alias{check_consec_na}
\title{Checks the longest sequence of consecutive missing values}
\usage{
check_consec_na(x, n)
}
\arguments{
\item{x}{Vector to check the proportion of missing values in.}

\item{n}{The threshold for the longest sequence of consecutive missing values
in \code{x}.}
}
\value{
\code{TRUE} if the longest sequence of consecutive missing values in
  \code{x} is less than or equal to \code{n} and \code{FALSE} otherwise.
}
\description{
\code{check_consec_na} returns \code{TRUE} if the longest sequence of
consecutive missing values is less than or equal to a given threshold. This
is a convenient wrapper function for \code{consec_na(x) <= n}.
}
\examples{
consec_na(c(1, NA, NA, NA, 2, NA, NA, 7), 4)

consec_na(c(rep(NA, 5), 1:2, rep(NA, 6)), 5)
}
