% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_checks.R
\name{check_n_na}
\alias{check_n_na}
\title{Check the number of missing values}
\usage{
check_n_na(x, n)
}
\arguments{
\item{x}{Vector to check the proportion of missing values in.}

\item{n}{The threshold for missing values in \code{x}.}
}
\value{
\code{TRUE} if the number of missing values in \code{x}
is less than or equal to \code{n} and \code{FALSE} otherwise.
}
\description{
\code{check_n_na} returns \code{TRUE} if the number of
missing values is less than or equal to a given threshold.
This is a convenient wrapper function for \code{n_na(x) <= n}.
}
\examples{
check_n_na(c(1, 2, NA, 4, NA, NA, 7), 5)

check_n_na(c(1:9, NA, NA, NA), 2)
}
