% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_checks.R
\name{check_n_non_na}
\alias{check_n_non_na}
\title{Check the number of non-missing values}
\usage{
check_n_non_na(x, n)
}
\arguments{
\item{x}{Vector to check the proportion of missing values in.}

\item{n}{The threshold for \strong{non-missing} values in \code{x}.}
}
\value{
\code{TRUE} if the number of \strong{non-missing} values in \code{x}
  is greater than or equal to \code{n} and \code{FALSE} otherwise.
}
\description{
\code{check_n_non_na} returns \code{TRUE} if the number of
\strong{non-missing} values is greater than or equal to a given threshold.
This is a convenient wrapper function for \code{n_non_na(x) >= n}.
}
\details{
Note that unlike the other \code{check_*} functions in \code{check_n_non_na}
the threshold is for a minimum number, since it is concerned with
\strong{non-missing} values.
}
\examples{
check_n_non_na(c(1, 2, NA, 4, NA, NA, 7), 5)

check_n_non_na(c(1:9, NA, NA, NA), 2)
}
