% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_calculations.R
\name{consec_na}
\alias{consec_na}
\title{Calculate the longest sequence of consecutive missing values}
\usage{
consec_na(x)
}
\arguments{
\item{x}{Vector to calculate the maximum number of consecutive missing values
in.}
}
\value{
The length of the longest sequence of consecutive missing values in
  \code{x}.
}
\description{
\code{consec_na} returns the length of the longest sequence of consecutive
missing values in \code{x}.
}
\details{
\code{consec_na} calculates the sequences of missing values in \code{x} and
returns the length of the longest sequence.
}
\examples{
consec_na(c(1, NA, NA, NA, 2, NA, NA, 7))

consec_na(c(rep(NA, 5), 1:2, rep(NA, 6)))
}
