% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_calculations.R
\name{n_na}
\alias{n_na}
\title{Calculate the number of missing values}
\usage{
n_na(x)
}
\arguments{
\item{x}{Vector to calculate the number of missing values in.}
}
\value{
The number of missing values in \code{x}.
}
\description{
\code{n_na} returns the number of missing values in \code{x}.
It is a convenient wrapper for \code{sum(is.na(x))}.
}
\examples{
n_na(c(1, 2, NA, 4, NA, NA, 7))

n_na(c(1:9, NA, NA))
}
