% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_calculations.R
\name{n_non_na}
\alias{n_non_na}
\title{Calculate the number of non-missing values}
\usage{
n_non_na(x)
}
\arguments{
\item{x}{Vector to calculate the number of \strong{non-missing} values in.}
}
\value{
The number of \strong{non-missing} values in \code{x}.
}
\description{
\code{n_non_na} returns the number of \strong{non-missing} values
in \code{x}.
It is a convenient wrapper for \code{sum(!is.na(x))}.
}
\examples{
n_non_na(c(1, 2, NA, 4, NA, NA, 7))

n_non_na(c(1:9, NA, NA))
}
