% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_check-individuals.R
\name{na_check_consec}
\alias{na_check_consec}
\title{Checks the longest sequence of consecutive missing values}
\usage{
na_check_consec(x, consec = NULL)
}
\arguments{
\item{x}{Vector to check the proportion of missing values in.}

\item{n}{The threshold for the longest sequence of consecutive missing values
in \code{x}.}
}
\value{
\code{TRUE} if the longest sequence of consecutive missing values in
  \code{x} is less than or equal to \code{n} and \code{FALSE} otherwise.
}
\description{
\code{na_check_consec} returns \code{TRUE} if the longest sequence of
consecutive missing values is less than or equal to a given threshold. This
is a convenient wrapper function for \code{na_consec(x) <= n}.
}
\examples{
na_check_consec(c(1, NA, NA, NA, 2, NA, NA, 7), 4)

na_check_consec(c(rep(NA, 5), 1:2, rep(NA, 6)), 5)
}
