% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_check-individuals.R
\name{na_check_n}
\alias{na_check_n}
\title{Check the number of missing values}
\usage{
na_check_n(x, n = NULL)
}
\arguments{
\item{x}{Vector to check the proportion of missing values in.}

\item{n}{The threshold for missing values in \code{x}.}
}
\value{
\code{TRUE} if the number of missing values in \code{x}
is less than or equal to \code{n} and \code{FALSE} otherwise.
}
\description{
\code{na_check_n} returns \code{TRUE} if the number of
missing values is less than or equal to a given threshold.
This is a convenient wrapper function for \code{na_n(x) <= n}.
}
\examples{
na_check_n(c(1, 2, NA, 4, NA, NA, 7), 5)

na_check_n(c(1:9, NA, NA, NA), 2)
}
