% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_na.R
\name{na_check_non_na}
\alias{na_check_non_na}
\title{Check the number of non-missing values}
\usage{
na_check_non_na(x, n_non = NULL)
}
\arguments{
\item{x}{Vector to check the proportion of missing values in.}

\item{n}{The threshold for \strong{non-missing} values in \code{x}.}
}
\value{
\code{TRUE} if the number of \strong{non-missing} values in \code{x}
  is greater than or equal to \code{n} and \code{FALSE} otherwise.
}
\description{
\code{na_check_non_na} returns \code{TRUE} if the number of
\strong{non-missing} values is greater than or equal to a given threshold.
This is a convenient wrapper function for \code{na_non_na(x) >= n}.
}
\details{
Note that unlike the other \code{check_*} functions in \code{na_check_non_na}
the threshold is for a minimum number, since it is concerned with
\strong{non-missing} values.
}
\examples{
na_check_non_na(c(1, 2, NA, 4, NA, NA, 7), 5)

na_check_non_na(c(1:9, NA, NA, NA), 2)
}
