% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_omit_if.R
\name{na_omit_if}
\alias{na_omit_if}
\title{Conditionally omit missing values}
\usage{
na_omit_if(x, prop, na, consec, na_non_na, prop_strict = FALSE)
}
\arguments{
\item{x}{Vector to omit missing values in if checks pass.}

\item{prop}{The proportion threshold (0 to 1) for missing values in \code{x}.
The proportion of missing values in \code{x} must be less than (or equal
to) \code{prop} for this check to pass, or strictly less than \code{prop}
if \code{strict = TRUE}.}

\item{na}{The threshold for the number of missing values in \code{x}. The
number of missing values in \code{x} must be less than or equal to \code{na} for this
check to pass.}

\item{consec}{The threshold for the longest sequence of consecutive missing
values in \code{x}. The longest sequence of consecutive missing values in \code{x} must
be less than or equal to \code{consec} for this check to pass.}

\item{na_non_na}{The threshold for \strong{non-missing} values in \code{x}. The
number of non-missing values in \code{x} must be greater than or equal to
\code{na_non_na} for this check to pass.}

\item{prop_strict}{A logical (default \code{FALSE}) indicating if the proportion
of missing values must be \strong{strictly} less than \code{prop}
(\code{strict = TRUE}) or only less than or equal to \code{na_prop}
(\code{strict = FALSE}). Ignored if \code{na_prop} is missing.}
}
\value{
A vector of the same type as \code{x}. Either \code{x} with missing
  values removed if all checks pass, or \code{x} unmodified if any checks
  fail.

  For consistency with \code{\link[stats::na.fail]{na.omit}}, if missing
  values are removed, the indices of the removed values form an
  \code{na.action} attribute of class \code{omit} in the result.

  If no missing values are removed (because the checks failed or there were
  no missing values in \code{x}) then no \code{na.action} attribute is added.
}
\description{
\code{na_omit_if} removes missing values from \code{x} if the specified checks
are satisfied, and returns \code{x} unmodified otherwise. When used within
summary functions, \code{na_omit_if} provides greater flexibility than
the \code{na.rm} option e.g. \code{sum(na_omit_if(x, prop = 0.05))}.
}
\details{
There are four type of checks available:
\itemize{
\item a maximum proportion of missing values allowed (\code{prop})
\item a maximum number of missing values allowed (\code{na})
\item a maximum number of consecutive missing values allowed (\code{consec}), and
\item a minimum number of non-missing values required (\code{na_non_na}).
}

Any number of checks may be specified, including none. If multiple checks are
specified, they must all be satisfied in order for missing values to be
omitted. If no checks are specified then missing values are omitted, since
this is considered as "all" checks passing. This is equivalent to \code{na.rm
= TRUE}.
}
\examples{
x <- c(1, 3, NA, NA, NA, 4, 2, NA, 4, 6)

sum(na_omit_if(x, prop = 0.45, na = 10, consec = 5))
sum(na_omit_if(x, prop = 0.45))

require(magrittr)
sum(x \%>\% na_omit_if(prop = 0.45))

# WMO specification for calculating monthly values from daily data
daily_rain <- rnorm(30)
daily_rain[c(3, 5, 6, 7, 8, 9, 24, 28)] <- NA
sum(daily_rain \%>\% na_omit_if(na = 10, consec = 4))
}
