% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_omit_if-individuals.R
\name{na_omit_if_prop}
\alias{na_omit_if_prop}
\alias{na_omit_if_n}
\alias{na_omit_if_consec}
\alias{na_omit_if_non_na}
\title{Conditionally omit missing values (single condition)}
\usage{
na_omit_if_prop(x, prop, strict = FALSE)

na_omit_if_n(x, n)

na_omit_if_consec(x, consec)

na_omit_if_non_na(x, n_non)
}
\arguments{
\item{x}{Vector to omit missing values in if checks pass.}

\item{prop}{The proportion threshold (0 to 1) for missing values in \code{x}.
The proportion of missing values in \code{x} must be less than (or equal
to) \code{prop} for this check to pass, or strictly less than \code{prop}
if \code{strict = TRUE}.}

\item{strict}{A logical (default \code{FALSE}) indicating if the proportion
of missing values must be \strong{strictly} less than \code{prop}
(\code{strict = TRUE}) or only less than or equal to \code{prop}
(\code{strict = FALSE}).}

\item{n}{The threshold for the number of missing values in \code{x}. The
number of missing values in \code{x} must be less than or equal to \code{n} for this
check to pass.}

\item{consec}{The threshold for the longest sequence of consecutive missing
values in \code{x}. The longest sequence of consecutive missing values in \code{x} must
be less than or equal to \code{consec} for this check to pass.}

\item{n_non}{The threshold for \strong{non-missing} values in \code{x}. The
number of non-missing values in \code{x} must be greater than or equal to
\code{n_non} for this check to pass.}
}
\description{
These set of functions remove missing values from \code{x} if the single,
specified check is satisfied, and returns \code{x} unmodified otherwise. They
are special cases of \code{\link{na_omit_if}}, which is the general case for
specifying multiple checks.
}
\details{
These functions replicate a subset of the functionality of
\code{\link{na_omit_if}}.

For example, \code{na_omit_if_consec(x, 4)} is equivalent to
\code{na_omit_if(x, consec = 4)}.

This more restricted form may be desirable when only a single check is
required.
}
