% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_calculations.R
\name{prop_na}
\alias{prop_na}
\title{Calculate the proportion of missing values}
\usage{
prop_na(x)
}
\arguments{
\item{x}{Vector to calculate the proportion of missing values in.}
}
\value{
The proportion (0 to 1) of missing values in \code{x}.
}
\description{
\code{prop_na} returns the proportion (0 to 1) of missing values in \code{x}.
It is a convenient wrapper for \code{mean(is.na(x))}.
}
\examples{
prop_na(c(1, 2, NA, 4))

prop_na(c(1:9, NA))
}
