% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_na.R
\name{na_check_prop}
\alias{na_check_prop}
\title{Check the proportion of missing values}
\usage{
na_check_prop(x, prop = NULL, strict = FALSE)
}
\arguments{
\item{x}{Vector to check the proportion of missing values in.}

\item{prop}{The threshold as a proportion (0 to 1) for missing values in
\code{x}.}

\item{strict}{A logical (default \code{FALSE}) indicating if the proportion
of missing values must be \strong{strictly} less than \code{prop}
(\code{strict = TRUE}) or only less than \code{na_prop} (\code{strict =
FALSE}).}
}
\value{
\code{TRUE} if the proportion of missing values in \code{x} is less
  than (or strictly less than if \code{strict = TRUE}) \code{prop}, and
  \code{FALSE} otherwise.
}
\description{
\code{na_check_prop} returns \code{TRUE} if the proportion of missing values
is below a given threshold. This is a convenient wrapper function for
\code{na_prop(x) < prop} or \code{na_prop(x) <= prop} (depending on
\code{strict}).
}
\details{
If \code{strict = FALSE} (the default), \code{na_check_prop} is equivalent to
\code{na_prop(x) <= prop} i.e. it returns \code{TRUE} if the proportion of
missing values in \code{x} is less than or equal to \code{na_prop}, and
\code{FALSE} otherwise.

If \code{strict = TRUE}, \code{na_check_prop} is equivalent to
\code{na_prop(x) < prop} i.e. the same as above but with a strictly less than
\code{prop} condition.
}
\examples{
na_check_prop(c(1, 2, NA, 4), 0.6)
na_check_prop(c(1, 2, NA, 4), 0.4)

na_check_prop(c(1:10, NA), 0.1)

na_check_prop(c(1:9, NA), 0.1, strict = TRUE)
}
